Move timestamp and emultiplex fixtures to plugin modules to clean conftest
